requirements.txt
fastapi>=0.115.0,<1.0.0
google-adk>=1.3.0,<2.0.0
google-cloud-firestore>=2.21.0,<3.0.0
google-cloud-storage>=2.19.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
uvicorn[standard]>=0.34.0,<1.0.0
